+++ szamlaagent/src/SzamlaAgent/Document/Document.php
@@ -24,6 +24,15 @@ class Document {
     const PAYMENT_METHOD_SZEP_CARD        = 'SZÉP kártya';
     const PAYMENT_METHOD_OTP_SIMPLE       = 'OTP Simple';
 
+    /** Lekérdezés számlaszám alapján */
+    const FROM_DOCUMENT_NUMBER = 1;
+
+    /** Lekérdezés rendelési szám alapján */
+    const FROM_ORDER_NUMBER = 2;
+
+    /** Számla lekérdezése külső számlaazonosító alapján */
+    const FROM_INVOICE_EXTERNAL_ID = 3;
+
     /**
      * Normál számla
      */