+++ szamlaagent/src/SzamlaAgent/SzamlaAgentUtil.php
@@ -212,6 +212,7 @@ class SzamlaAgentUtil {
      * @return array
      */
     public static function checkValidXml($xmlContent) {
+        libxml_clear_errors();
         libxml_use_internal_errors(true);
 
         $doc = new \DOMDocument('1.0', 'utf-8');